Added AUR installation when using arch support

+++ install-artix.sh
@@ -315,24 +315,29 @@ fi
 
 # Install AUR helper
 if [[ $enable_aur == true ]]; then
-    # install dependencies
-    deps="git pacutils perl-{libwww,term-ui,json,data-dump,lwp-protocol-https,term-readline-gnu}"
-    artix-chroot /mnt bash -c "pacman --noconfirm --needed -Sy ${deps}"
+    if [[ $arch_support == true ]]; then
+        artix-chroot /mnt bash -c "pacman --noconfirm -Syy trizen"
+    else
+        # install dependencies
+        deps="git pacutils perl-{libwww,term-ui,json,data-dump,lwp-protocol-https,term-readline-gnu}"
+        artix-chroot /mnt bash -c "pacman --noconfirm --needed -Sy ${deps}"
 
-    # download package
-    rm -rf /mnt/home/${user}/trizen &> /dev/null
-    clone="git clone https://aur.archlinux.org/trizen"
-    artix-chroot /mnt bash -c "runuser -l ${user} -c \"${clone}\""
+        # download package
+        rm -rf /mnt/home/${user}/trizen &> /dev/null
+        clone="git clone https://aur.archlinux.org/trizen"
+        artix-chroot /mnt bash -c "runuser -l ${user} -c \"${clone}\""
 
-    # build package
-    build="cd /home/${user}/trizen && makepkg --noconfirm"
-    artix-chroot /mnt bash -c "runuser -l ${user} -c \"${build}\""
+        # build package
+        build="cd /home/${user}/trizen && makepkg --noconfirm"
+        artix-chroot /mnt bash -c "runuser -l ${user} -c \"${build}\""
 
-    # add package to pacman
-    artix-chroot /mnt bash -c "pacman -U --noconfirm /home/${user}/trizen/*.zst"
+        # add package to pacman
+        artix-chroot /mnt bash -c "pacman -U --noconfirm /home/${user}/trizen/*.zst"
 
-    # remove unnecessary files
-    rm -rf "/mnt/home/${user}/trizen"
+        # remove unnecessary files
+        rm -rf "/mnt/home/${user}/trizen"
+    fi
+    echo "AUR helper installation complete!"
 fi
 
 # FINISH
